Export intent removed from GUI attributes
GUI no longer has an export mode built in. It has been reworked so that export intent can be supplied to the functions and not inferred from the state of the GUI itself. This means there doesn't need to be a special alternative preamble to MRM files indicating export intent.