Implemented autoclose
From the Command Line / Terminal, Morpho can now take an optional flag "--autoclose" which will automatically close the animation window when the animation finishes. This is only used for the initial "show-off" animation, which will auto close and bring the user directly to the GUI without having to close the animation window first.