Minor comment change
I have a better explanation for the laggy successive animation plays.
I think I could fix it for real now, but having Morpho run a copy of itself to play an animation is already so ingrained into the code that I don't want to change it. Anyway, the way it works now helps guard against any other potential memory clean-up issues (if there are any).